refactor: replace and flesh out discussions data layer
Since we only have one GET endpoint for all the data around discussions, we don’t need two separate slices/thunks/api files.  The API needs to be hit once when the discussions settings load.  This means the app-list and app-config-form share far more state than originally thought.

The AppList and AppConfigForm are no longer responsible for data loading - we’ve moved that back up into DiscussionsSettings.  Now they just read from the redux state and load what they need.

The main change is moving app-list/data/slice.js up to discussions/data/slice.js, renaming the reducer, and adding a saveStatus to it - turns out this is all we need.  The original two slices go away.

The discussions/data/api file gets a proper implementation of postAppConfig which mostly works with the server as of this writing - we have some data shape issues to figure out.

AppConfigForm needed a little help.  It now checks whether our data is loaded and selects an app based on its route.  This allows us to deep link in and keep the selectedAppId correct.